refactor(phase-3B-1): extract _Handler from server.py into routes.py [no behavior change]
Phase 3B PR-3B-1 / viz/REFACTORING.md §2 Step 1 — final viz split.

``ari/viz/server.py`` shrinks from 1336 → 201 lines.  The HTTP request
handler class ``_Handler`` (do_GET / do_POST dispatch chain, ~1130 L)
plus the ``_write_access_log`` helper move to a new
``ari/viz/routes.py``.

The dispatch chain is preserved verbatim:
- if-elif chain order unchanged.
- ``_REDACT_KEYS`` literal still owned by ``ui_helpers.py``.
- ``_access_log_lock`` moved alongside its only consumer.

server.py keeps:
- ``_DualStackServer`` (IPv6/IPv4 dual-stack ThreadingHTTPServer).
- ``_http_thread`` / ``_main`` / ``main`` entry points.
- Re-exports of ``_Handler`` / ``_write_access_log`` so existing
  ``from ari.viz.server import ...`` paths keep working.

Test fixture updates (cross-file source-text grep tests):
- ``test_settings_roundtrip._srv()`` now also concatenates routes.py.
- New ``_viz_server_concat(viz_dir)`` helper added to:
  - test_server.py
  - test_wizard.py
  - test_api_lineage_decisions.py
  - test_data_flow.py
  - test_default_provider.py
  - test_file_explorer.py
  - test_launch_config.py
  - test_page_requirements.py
  Each helper concats ``ui_helpers.py + websocket.py + routes.py +
  server.py`` so source-text checks find the moved literals.

pytest ari-core/tests/ -q (excluding live integrations):
1971 passed, 15 skipped (unchanged).
